+++ src/shinestacker/algorithms/base_stack_algo.py
@@ -2,6 +2,7 @@
 import numpy as np
 from .. core.exceptions import InvalidOptionError, ImageLoadError
 from .. config.constants import constants
+from .. core.colors import color_str
 from .utils import read_img, get_img_metadata, validate_image
 
 
@@ -27,7 +28,7 @@ class BaseStackAlgo:
         return self._steps_per_frame
 
     def print_message(self, msg):
-        self.process.sub_message_r(msg)
+        self.process.sub_message_r(color_str(msg, constants.LOG_COLOR_LEVEL_3))
 
     def read_image_and_update_metadata(self, img_path, metadata):
         img = read_img(img_path)