+++ src/shinestacker/algorithms/multilayer.py
@@ -176,7 +176,9 @@ class MultiLayer(JobBase, FrameMultiDirectory):
         else:
             raise RuntimeError("input_path option must contain a path or an array of paths")
         if len(paths) == 0:
-            self.print_message(color_str("no input paths specified", "red"), level=logging.WARNING)
+            self.print_message(color_str("no input paths specified",
+                                         constants.LOG_COLOR_LEVEL_ALERT),
+                               level=logging.WARNING)
             return
         files = self.folder_filelist()
         if len(files) == 0:
@@ -184,22 +186,23 @@ class MultiLayer(JobBase, FrameMultiDirectory):
                 color_str(f"no input in {len(paths)} specified path" +
                           ('s' if len(paths) > 1 else '') + ": "
                           ", ".join([f"'{p}'" for p in paths]),
-                          "red"),
+                          constants.LOG_COLOR_LEVEL_ALERT),
                 level=logging.WARNING)
             return
-        self.print_message(color_str("merging frames in " + self.folder_list_str(), "blue"))
+        self.print_message(color_str("merging frames in " + self.folder_list_str(),
+                           constants.LOG_COLOR_LEVEL_2))
         input_files = [f"{self.working_path}/{f}" for f in files]
         self.print_message(
-            color_str("frames: " + ", ".join([i.split("/")[-1] for i in files]), "blue"))
-        self.print_message(
-            color_str("reading files", "blue"))
+            color_str("frames: " + ", ".join([i.split("/")[-1] for i in files]),
+                      constants.LOG_COLOR_LEVEL_2))
+        self.print_message(color_str("reading files", constants.LOG_COLOR_LEVEL_2))
         filename = ".".join(files[0].split("/")[-1].split(".")[:-1])
         output_file = f"{self.working_path}/{self.output_path}/{filename}.tif"
         callbacks = {
             'exif_msg': lambda path: self.print_message(
-                color_str(f"copying exif data from path: {path}", "blue")),
+                color_str(f"copying exif data from path: {path}", constants.LOG_COLOR_LEVEL_2)),
             'write_msg': lambda path: self.print_message(
-                color_str(f"writing multilayer tiff file: {path}", "blue"))
+                color_str(f"writing multilayer tiff file: {path}", constants.LOG_COLOR_LEVEL_2))
         }
         write_multilayer_tiff(input_files, output_file, labels=None, exif_path=self.exif_path,
                               callbacks=callbacks)